Added function for verifying users only on username instead of username and password wich didn't work
Replaced chosen library with angular-ui-select. The angular-ui-select library does work on mobile. AddUser and EditUser have been modified to accomendate the new library.

Changed the role type in the database. Now its saved as INT instead of VarChar. Values are saved as 0 or 1. Now the role selection in the pages do work as intended.